refactor : validator => validate
