+++ src/validations/validate/ReStartValidate.js
@@ -1,6 +1,11 @@
-import { Y_OR_NO_ERROR_MESSAGE } from "../../constants/constants.js";
-import { reStartValidator } from "../validator/reStartValidator.js";
+import { Y_OR_NO_ERROR_MESSAGE, YES, NO } from "../../constants/constants.js";
 import runValidators from "../../utils/runValidators.js";
+
+export const reStartValidator = {
+  isYesOrNo(input){
+    return input===YES||input===NO;
+  },
+};
 
 const validateYorN = (input) => {
   if (!reStartValidator.isYesOrNo(input)) {
