--- src/validations/validator/BonusNumberValidator.js
@@ -1,5 +1,0 @@
-export const bonusNumberValidator = {
-  isDuplicated(winningNumbers, bonusNumber) {
-    return winningNumbers.includes(bonusNumber);
- },
-};
