--- src/validations/validator/PurchaseMoneyValidator.js
@@ -1,11 +1,0 @@
-import { LOTTO_CONDITION } from '../../constants/constants.js';
-
-export const purchaseMoneyValidator = {
-  isValidUnit(input) {
-    return input % LOTTO_CONDITION.PRICE === 0;
-  },
-
-  isValidRange(input) {
-    return LOTTO_CONDITION.PRICE <= input;
-  },
-};
