--- src/validations/validator/ReStartValidator.js
@@ -1,9 +1,0 @@
-import { YES } from "../../constants/constants.js";
-import { NO } from "../../constants/constants.js";
-
-
-export const reStartValidator = {
-  isYesOrNo(input){
-    return input===YES||input===NO;
-  },
-};
